Mark more operations as pure
