+++ ecmascript/minifier/src/compress/optimize/conditionals.rs
@@ -1,11 +1,7 @@
 use super::Optimizer;
-use crate::compress::optimize::Ctx;
 use crate::compress::optimize::DISABLE_BUGGY_PASSES;
-use crate::compress::util::negate_cost;
-use crate::debug::dump;
 use crate::util::make_bool;
 use crate::util::SpanExt;
-use std::mem::swap;
 use swc_common::EqIgnoreSpan;
 use swc_common::Spanned;
 use swc_common::DUMMY_SP;
@@ -23,70 +19,6 @@ use swc_ecma_utils::Value::Known;
 /// Methods related to the option `conditionals`. All methods are noop if
 /// `conditionals` is false.
 impl Optimizer<'_> {
-    pub(super) fn negate_cond_expr(&mut self, cond: &mut CondExpr) {
-        if negate_cost(&cond.test, true, false).unwrap_or(isize::MAX) >= 0 {
-            return;
-        }
-
-        self.changed = true;
-        log::debug!("conditionals: `a ? foo : bar` => `!a ? bar : foo` (considered cost)");
-        let start_str = dump(&*cond);
-
-        {
-            let ctx = Ctx {
-                in_bool_ctx: true,
-                ..self.ctx
-            };
-            self.with_ctx(ctx).negate(&mut cond.test);
-        }
-        swap(&mut cond.cons, &mut cond.alt);
-
-        if cfg!(feature = "debug") {
-            log::trace!(
-                "[Change] Negated cond: `{}` => `{}`",
-                start_str,
-                dump(&*cond)
-            )
-        }
-    }
-
-    /// Negates the condition of a `if` statement to reduce body size.
-    pub(super) fn negate_if_stmt(&mut self, stmt: &mut IfStmt) {
-        let alt = match stmt.alt.as_deref_mut() {
-            Some(v) => v,
-            _ => return,
-        };
-
-        match &*stmt.cons {
-            Stmt::Return(..) | Stmt::Continue(ContinueStmt { label: None, .. }) => return,
-            _ => {}
-        }
-
-        if negate_cost(&stmt.test, true, false).unwrap_or(isize::MAX) < 0 {
-            self.changed = true;
-            log::debug!("if_return: Negating `cond` of an if statement which has cons and alt");
-            let ctx = Ctx {
-                in_bool_ctx: true,
-                ..self.ctx
-            };
-            self.with_ctx(ctx).negate(&mut stmt.test);
-            swap(alt, &mut *stmt.cons);
-            return;
-        }
-
-        match &*alt {
-            Stmt::Return(..) | Stmt::Continue(ContinueStmt { label: None, .. }) => {
-                self.changed = true;
-                log::debug!(
-                    "if_return: Negating an if statement because the alt is return / continue"
-                );
-                self.negate(&mut stmt.test);
-                swap(alt, &mut *stmt.cons);
-            }
-            _ => return,
-        }
-    }
-
     /// This method may change return value.
     ///
     /// - `a ? b : false` => `a && b`
